requirements.txt
torch
torchvision
matplotlib
numpy
fastapi
uvicorn
pillow
python-multipart